parse_silkworm_command_line
