+++ cmd/silkworm.cpp
@@ -67,6 +67,174 @@ class ResourceUsageLog : public ActiveComponent {
     }
 };
 
+void parse_silkworm_command_line(CLI::App& cli, int argc, char* argv[], SilkwormCoreSettings& settings) {
+    using namespace silkworm::cmd;
+
+    auto& node_settings = settings.node_settings;
+
+    // Node settings
+    std::filesystem::path data_dir_path;
+    std::string chaindata_max_size_str{human_size(node_settings.chaindata_env_config.max_size)};
+    std::string chaindata_growth_size_str{human_size(node_settings.chaindata_env_config.growth_size)};
+    std::string chaindata_page_size_str{human_size(node_settings.chaindata_env_config.page_size)};
+    std::string batch_size_str{human_size(node_settings.batch_size)};
+    std::string etl_buffer_size_str{human_size(node_settings.etl_buffer_size)};
+    add_option_data_dir(cli, data_dir_path);
+
+    cli.add_flag("--chaindata.exclusive", node_settings.chaindata_env_config.exclusive,
+                 "Chaindata database opened in exclusive mode");
+    cli.add_flag("--chaindata.readahead", node_settings.chaindata_env_config.read_ahead,
+                 "Chaindata database enable readahead");
+    cli.add_flag("--chaindata.writemap", node_settings.chaindata_env_config.write_map,
+                 "Chaindata database enable writemap");
+
+    cli.add_option("--chaindata.growthsize", chaindata_growth_size_str, "Chaindata database growth size.")
+        ->capture_default_str()
+        ->check(HumanSizeParserValidator("64MB"));
+    cli.add_option("--chaindata.pagesize", chaindata_page_size_str, "Chaindata database page size. A power of 2")
+        ->capture_default_str()
+        ->check(HumanSizeParserValidator("256B", {"65KB"}));
+    cli.add_option("--chaindata.maxsize", chaindata_max_size_str, "Chaindata database max size.")
+        ->capture_default_str()
+        ->check(HumanSizeParserValidator("32MB", {"128TB"}));
+
+    cli.add_option("--batchsize", batch_size_str, "Batch size for stage execution")
+        ->capture_default_str()
+        ->check(HumanSizeParserValidator("64MB", {"16GB"}));
+    cli.add_option("--etl.buffersize", etl_buffer_size_str, "Buffer size for ETL operations")
+        ->capture_default_str()
+        ->check(HumanSizeParserValidator("64MB", {"1GB"}));
+
+    add_option_private_api_address(cli, node_settings.private_api_addr);
+
+    // Sentry settings
+    add_option_external_sentry_address(cli, node_settings.external_sentry_addr);
+
+    add_option_sentry_api_address(cli, node_settings.sentry_api_addr);
+
+    cli.add_option("--sync.loop.throttle", node_settings.sync_loop_throttle_seconds,
+                   "Sets the minimum delay between sync loop starts (in seconds)")
+        ->capture_default_str()
+        ->check(CLI::Range(1u, 7200u));
+
+    cli.add_option("--sync.loop.log.interval", node_settings.sync_loop_log_interval_seconds,
+                   "Sets the interval between sync loop logs (in seconds)")
+        ->capture_default_str()
+        ->check(CLI::Range(10u, 600u));
+
+    cli.add_flag("--fakepow", node_settings.fake_pow, "Disables proof-of-work verification");
+
+    // Chain options
+    add_option_chain(cli, node_settings.network_id);
+
+    // Prune options
+    std::string prune_mode;
+    auto& prune_opts = *cli.add_option_group("Prune", "Prune options to delete ancient data from DB");
+    prune_opts
+        .add_option("--prune", prune_mode,
+                    "Delete data older than 90K blocks (see \"--prune.*.older\" for different height)\n"
+                    "h - prune history (ChangeSets, HistoryIndices - used by historical state access)\n"
+                    "r - prune receipts (Receipts, Logs, LogTopicIndex, LogAddressIndex - used by eth_getLogs and "
+                    "similar RPC methods)\n"
+                    "s - prune senders recovered\n"
+                    "t - prune transaction by it's hash index\n"
+                    "c - prune call traces (used by trace_* methods)\n"
+                    "If item is NOT in the list - means NO pruning for this data.\n"
+                    "Example: --prune=hrtc (default: none)")
+        ->capture_default_str()
+        ->check(PruneModeValidator());
+
+    prune_opts.add_option("--prune.h.older", "Override default 90k blocks of history to prune")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.r.older", "Override default 90k blocks of receipts to prune")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.s.older", "Override default 90k blocks of senders to prune")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.t.older", "Override default 90k blocks of transactions to prune")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.c.older", "Override default 90k blocks of call traces to prune")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.h.before", "Prune history data before this block")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.r.before", "Prune receipts data before this block")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.s.before", "Prune senders data before this block")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.t.before", "Prune transactions data before this block")
+        ->check(CLI::Range(0u, UINT32_MAX));
+    prune_opts.add_option("--prune.c.before", "Prune call traces data before this block")
+        ->check(CLI::Range(0u, UINT32_MAX));
+
+    // Logging options
+    auto& log_settings = settings.log_settings;
+    add_logging_options(cli, log_settings);
+
+    // RPC server options
+    auto& server_settings = settings.server_settings;
+
+    uint32_t num_contexts;
+    add_option_num_contexts(cli, num_contexts);
+
+    silkworm::rpc::WaitMode wait_mode;
+    add_option_wait_mode(cli, wait_mode);
+
+    // Snapshot&Bittorrent options
+    auto& snapshot_settings = settings.snapshot_settings;
+    add_snapshot_options(cli, snapshot_settings);
+
+    cli.parse(argc, argv);
+
+    // Validate and assign settings
+    const auto chaindata_page_size{parse_size(chaindata_page_size_str)};
+    if (!chaindata_page_size.has_value() || (*chaindata_page_size & (*chaindata_page_size - 1)) != 0) {
+        throw std::invalid_argument("--chaindata.pagesize is not a power of 2");
+    }
+    node_settings.chaindata_env_config.page_size = chaindata_page_size.value();
+    const auto mdbx_max_size_hard_limit{chaindata_page_size.value() * db::kMdbxMaxPages};
+    const auto chaindata_max_size{parse_size(chaindata_max_size_str)};
+    if (chaindata_max_size.value() > mdbx_max_size_hard_limit) {
+        throw std::invalid_argument("--chaindata.maxsize exceeds max allowed size by page size i.e" +
+                                    human_size(mdbx_max_size_hard_limit));
+    }
+    const auto chaindata_growth_size{parse_size(chaindata_growth_size_str)};
+    if (chaindata_growth_size > (mdbx_max_size_hard_limit / /* two increments ?*/ 2u)) {
+        throw std::invalid_argument("--chaindata.growthsize must be <=" + human_size(mdbx_max_size_hard_limit / 2));
+    }
+
+    node_settings.data_directory = std::make_unique<DataDirectory>(data_dir_path, /*create=*/true);
+    node_settings.chaindata_env_config.max_size = chaindata_max_size.value();
+    node_settings.chaindata_env_config.growth_size = chaindata_growth_size.value();
+
+    node_settings.batch_size = parse_size(batch_size_str).value();
+    node_settings.etl_buffer_size = parse_size(etl_buffer_size_str).value();
+
+    // Parse prune mode
+    db::PruneDistance olderHistory, olderReceipts, olderSenders, olderTxIndex, olderCallTraces;
+    if (cli["--prune.h.older"]->count()) olderHistory.emplace(cli["--prune.h.older"]->as<BlockNum>());
+    if (cli["--prune.r.older"]->count()) olderReceipts.emplace(cli["--prune.r.older"]->as<BlockNum>());
+    if (cli["--prune.s.older"]->count()) olderSenders.emplace(cli["--prune.s.older"]->as<BlockNum>());
+    if (cli["--prune.t.older"]->count()) olderTxIndex.emplace(cli["--prune.t.older"]->as<BlockNum>());
+    if (cli["--prune.c.older"]->count()) olderCallTraces.emplace(cli["--prune.c.older"]->as<BlockNum>());
+
+    db::PruneThreshold beforeHistory, beforeReceipts, beforeSenders, beforeTxIndex, beforeCallTraces;
+    if (cli["--prune.h.before"]->count()) beforeHistory.emplace(cli["--prune.h.before"]->as<BlockNum>());
+    if (cli["--prune.r.before"]->count()) beforeReceipts.emplace(cli["--prune.r.before"]->as<BlockNum>());
+    if (cli["--prune.s.before"]->count()) beforeSenders.emplace(cli["--prune.s.before"]->as<BlockNum>());
+    if (cli["--prune.t.before"]->count()) beforeTxIndex.emplace(cli["--prune.t.before"]->as<BlockNum>());
+    if (cli["--prune.c.before"]->count()) beforeCallTraces.emplace(cli["--prune.c.before"]->as<BlockNum>());
+
+    node_settings.prune_mode =
+        db::parse_prune_mode(prune_mode,  //
+                             olderHistory, olderReceipts, olderSenders, olderTxIndex, olderCallTraces, beforeHistory,
+                             beforeReceipts, beforeSenders, beforeTxIndex, beforeCallTraces);
+
+    server_settings.set_address_uri(node_settings.private_api_addr);
+    server_settings.set_num_contexts(num_contexts);
+    server_settings.set_wait_mode(wait_mode);
+
+    snapshot_settings.bittorrent_settings.repository_path = snapshot_settings.repository_dir;
+}
+
 // main
 int main(int argc, char* argv[]) {
     using namespace boost::placeholders;
